+++ packages/Webkul/Admin/src/Http/Controllers/Catalog/AttributeFamilyController.php
@@ -1,6 +1,6 @@
 <?php
 
-namespace Webkul\Admin\Http\Controllers\Attribute;
+namespace Webkul\Admin\Http\Controllers\Catalog;
 
 use Illuminate\Support\Facades\Event;
 use Webkul\Admin\Http\Controllers\Controller;
